server/requirements.txt
alembic==0.9.8
flask==1.0.3
Flask-SQLAlchemy==2.3.2
SQLAlchemy==1.2.5
Flask-Migrate==2.5.1
psycopg2==2.8.2
jsonschema==3.0.1
Flask-JWT-Extended==3.18.2
flask-cors==3.0.7
python-magic==0.4.15
gunicorn==19.9.0
pytest==4.5.0
pytest-cov==2.7.1
boto3==1.9.157
